feat: refactor logic as useOnceInchRate hook
